backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Google DeepMind Hackathon 2026 - Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "google-genai>=1.0.0",
    "psycopg2-binary>=2.9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]
